Fixed package name and project refactoring
As a standard, package names should be in all lowercase and it is
recommended to use a reversed internet domain. I suggest
com.jadventure.game. I also tried to refactor the project folder
structure to a more recognized one. I added .classpath and .project
files for people using eclipse so that it is easier to open project.
This will also help people importing the project to IntelliJ or Netbeans