Menus major changes
So made some major changes on this commit.
Menus is no longer holding the actual menus to be used, any menus now
will need their own class and the values will be held there.

Created MainMenu.java and PlayerMenu.java
Decided to go with this approach so that when the menu is done we can
null it.
Also since Player is copied to the submenu every time it is called, we
know we will be working with current values of the Player object.
Assigned menu ID's to each sub-menu Menus.determineMenu will know which
one to pull.

Have tested current menus so far and they seem to be working.