Added classes for game navigation
Added classes needed for game navigation. LocationManager is a singleton
that loads all locations from a json file. Player has now an associated
location which is saved with his profile. To finish navigation, we have
to implement a command parser for normal gameplay. Menus are good, but
not for normal gameplay. We have to be able to interpret commands given
by the player, such as "go north", "north", "attack bear", etc.